refactor: refactor match scoresheet component and related dialogs for match management
- Introduced the Scoresheet component to display match details and scores.
- Added dialogs for managing rounds, comments, and player/team edits.
- Implemented functionality for manual winner selection and tie-breaking.
- Enhanced image handling for match visuals with upload and delete capabilities.
- Refactored match data handling to improve user experience and data consistency.
