+++ apps/nextjs/src/components/match/scoresheet/DetailDialog.tsx
